Fix CRC Greentea test
As app.cmake sets the programming languages to use for the project, including the `project` command in CMake before including `app.cmake` causes an error as the programming language are somehow set by that command.

The `project` command has been moved further down in the CMakeLists.txt similarly to other tests.

Signed-off-by: Hugues Kamba-Mpiana <hugues.kambampiana@arm.com>